Add more OOP, create new classes
