+++ data/duke.txt
@@ -1,5 +1,6 @@
-D|1|eat|2023-12-12 11:12
-D|0|eat|2025-12-12 11:12
-D|0|eat|2023-12-12 11:18
 E|0|eat|2023-12-12 11:12|2023-12-12 11:15
 E|0|sleep|2023-12-12 11:12|2023-12-12 13:15
+T|0|add
+T|0|add
+E|0|add|2100-12-12 12:12|2100-12-13 12:15
+T|0|hello
